Fixed default prediction method.
The default method, described in Fan & Ramamohanarao (2006) is score(T,Ci) = sum(suppCi(e)) wich e is an instance that covers the example.

Now, for each class, get the normalised score, to do that, divide score(T, Ci) by the median value of all supports of tha patterns that covers the example T for class Ci.