23/03/2017
Changing the way create the filters. We are trying that the user can specify the filters that he want to apply directly. Thus, there it is necessary to change the base of the framework due to to possible of the application of n filters instead of a fixed number (3 in the last version).

- We change now the method saveMeasures() that saves the results in files.
- Single execution has been modified to the new schema of filtering (first verions of this schema)
- It is necessary to change the predict() method of all algorithms.
- It is necessary to adapt the k-fold execution to this new behavior.
- Also, it is necessary to adapt the command-line version of those functions.

 Finally, it is necessary to improve functions in order to remove duplicities !!!