24/03/2017
Modified the framework to work with any number of filters. We change the base predict() method of Model to return only one set of predictions by set of patterns. Also, we centralise the filter and prediction behaviours of the method in order to remove duplicities in batch and in command line modes.